Moved boltdb implementation into a seperate file.
